+++ public/sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'movilidad-v1.27.1'
+const CACHE_NAME = 'movilidad-v1.27.2'
 
 self.addEventListener('install', (event) => {
   event.waitUntil(